better installer
git-svn-id: https://modzer0.cs.uaf.edu/svn/bcode/trunk/yum@220 d3944efe-84e1-0310-922a-b45035249031

+++ fc3/install.sh
@@ -1,11 +1,26 @@
 #!/bin/sh
 
-echo "moving /etc/yum.conf to /etc/yum.conf.default"
-mv /etc/yum.conf /etc/yum.conf.default
+if [ -f ./yum.conf ] ; then
+	if [ -f /etc/yum.conf.default ] ; then
+		echo "/etc/yum.conf.default detected"
+		echo "aborting.... clean this up yourself"
+		exit
+	fi
+	echo "moving /etc/yum.conf to /etc/yum.conf.default"
+	mv /etc/yum.conf /etc/yum.conf.default
+	echo "Installing cleaner yum.conf"
+	cp yum.conf /etc/yum.conf
+fi
+
+if [ -d /etc/yum.repos.default ] ; then
+        echo "/etc/yum.repos.default detected"
+        echo "aborting.... clean this up yourself"
+        exit;
+fi
+
 echo "moving /etc/yum.repos.d to /etc/yum.repos.default"
 mv /etc/yum.repos.d /etc/yum.repos.default
-echo "Installing cleaner yum.conf"
-cp yum.conf /etc/yum.conf
+
 echo "Putting version tracked yum.repos.d in place"
 cp -r . /etc/yum.repos.d
 
@@ -41,4 +56,3 @@ echo "you can stay in sync w/ bishop's yum.repos.d by going to"
 echo "/etc/yum.repos.d and typing: svn update"
 echo
 echo "Enjoy! .bishop"
-
